fix dtype issue

+++ ageUpscaling/upscaling/upscaling.py
@@ -320,14 +320,14 @@ class UpscaleAge(ABC):
                                                                         "longitude": subset_features_cube.longitude,
                                                                         "time": subset_features_cube.time.values,                                                          
                                                                         'members': [run_]}, 
-                                                                dims=["latitude", "longitude", "time", "members"]).astype('int16')}
+                                                                dims=["latitude", "longitude", "time", "members"])}
                     if self.upscaling_config['fuse_wLandsat']:
                         output_data["forest_age_hybrid"] = xr.DataArray(fused_pred_age, 
                                                                        coords={"latitude": subset_features_cube.latitude, 
                                                                                "longitude": subset_features_cube.longitude,
                                                                                "time": subset_features_cube.time.values,                                                          
                                                                                'members': [run_]}, 
-                                                                       dims=["latitude", "longitude", "time", "members"]).astype('int16')
+                                                                       dims=["latitude", "longitude", "time", "members"])
                     
                     output_reg_xr.append(xr.Dataset(output_data))
